macros: Move utility things into util module
